moved memory related utils into their own file to avoid circular import with System
