+++ openmdao/devtools/debug.py
@@ -272,59 +272,3 @@ def config_summary(problem, stream=sys.stdout):
     print("Nonlinear Solvers: %s" % sorted(set(type(s.nonlinear_solver).__name__ for s in
                                                          allgroups)),
           file=stream)
-
-try:
-    import resource
-
-    def max_mem_usage():
-        """
-        Returns
-        -------
-        The max memory used by this process and its children, in MB.
-        """
-        denom = 1024.
-        if sys.platform == 'darwin':
-            denom *= denom
-        total = resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / denom
-        total += resource.getrusage(resource.RUSAGE_CHILDREN).ru_maxrss / denom
-        return total
-except ImportError:
-    resource = None
-
-try:
-    import psutil
-
-    def mem_usage(msg='', out=sys.stdout):
-        """
-        Returns
-        -------
-        The current memory used by this process (and it's children?), in MB.
-        """
-        denom = 1024. * 1024.
-        p = psutil.Process(os.getpid())
-        mem = p.memory_info().rss / denom
-        if msg:
-            print(msg,"%6.3f MB" % mem, file=out)
-        return mem
-
-    def diff_mem(fn):
-        """
-        This gives the difference in memory before and after the
-        decorated function is called. Requires psutil to be installed.
-        """
-        @functools.wraps(fn)
-        def wrapper(*args, **kwargs):
-            startmem = mem_usage()
-            ret = fn(*args, **kwargs)
-            maxmem = mem_usage()
-            diff = maxmem - startmem
-            if diff > 0.0:
-                if args and hasattr(args[0], 'pathname'):
-                    name = args[0].pathname
-                else:
-                    name = ''
-                print(name,"%s added %5.3f MB (total: %6.3f)" % (fn.__name__, diff, maxmem))
-            return ret
-        return wrapper
-except ImportError:
-    mem_usage = diff_mem = None
